Last ver.

+++ src/main/java/com/shrralis/ssblog/servlet/post/DeletePostServlet.java
@@ -1,7 +1,6 @@
 package com.shrralis.ssblog.servlet.post;
 
 import com.shrralis.ssblog.dto.DeletePostDTO;
-import com.shrralis.ssblog.entity.User;
 import com.shrralis.ssblog.service.PostServiceImpl;
 import com.shrralis.ssblog.service.interfaces.IPostService;
 import com.shrralis.ssblog.servlet.base.ServletWithGsonProcessor;
@@ -15,7 +14,6 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.net.URLDecoder;
 import java.sql.SQLException;
 
 @WebServlet("/deletePost")
@@ -24,31 +22,33 @@ public class DeletePostServlet extends ServletWithGsonProcessor {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        IPostService service;
-        DeletePostDTO dto = new DeletePostDTO.Builder()
-                .setCookieUser(getGson().fromJson(
-                        URLDecoder.decode(req.getSession(false).getAttribute("user").toString(), "UTF-8"),
-                        User.class
-                ))
-                .setPostId(Integer.valueOf(req.getParameter("id")))
-                .build();
-
-        try {
-            service = new PostServiceImpl();
-        } catch (ClassNotFoundException | SQLException e) {
-            logger.debug("Exception!", e);
-            req.setAttribute("error", JsonError.Error.UNEXPECTED.getMessage());
-            resp.sendRedirect("/");
+        if (req.getParameter("confirm") == null) {
+            req.setAttribute("id", req.getParameter("id"));
+            getServletContext().getRequestDispatcher("/deletePost.jsp").forward(req, resp);
             return;
-        }
+        } else if (req.getParameter("confirm").equalsIgnoreCase("true")) {
+            IPostService service;
+            DeletePostDTO dto = DeletePostDTO.Builder.aDeletePostDTO()
+                    .setCookieUser(getCookieUser(req))
+                    .setPostId(Integer.valueOf(req.getParameter("id")))
+                    .build();
 
-        JsonResponse response = service.delete(dto);
+            try {
+                service = new PostServiceImpl();
+            } catch (ClassNotFoundException | SQLException e) {
+                logger.debug("Exception!", e);
+                req.setAttribute("error", JsonError.Error.UNEXPECTED.getMessage());
+                resp.sendRedirect("/");
+                return;
+            }
 
-        if (response.getResult().equals(JsonResponse.OK)) {
-            resp.sendRedirect("/");
-        } else {
-            req.setAttribute("error", JsonError.Error.UNEXPECTED.getMessage());
-            resp.sendRedirect("/");
+            JsonResponse response = service.delete(dto);
+
+            if (!response.getResult().equals(JsonResponse.OK)) {
+                req.setAttribute("error", JsonError.Error.UNEXPECTED.getMessage());
+                return;
+            }
         }
+        resp.sendRedirect("/");
     }
 }
